pyproject.toml
[tool.poetry]
name = "registry-cli"
version = "0.1.0"
description = "CLI tool for managing student records"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
sqlalchemy = "^2.0.23"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
selenium = "^4.15.2"
lxml = "^4.9.3"
reportlab = "^4.0.8"
python-dotenv = "^1.0.1"
libsql-client = "^0.3.1"
sqlalchemy-libsql = "^0.1.0"
openpyxl = "^3.1.5"
nanoid = "^2.0.0"
pypdf2 = "^3.0.1"

[tool.poetry.scripts]
registry = "registry_cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
